Lib/glob.py
"""Filename globbing utility."""

zaimportuj os
zaimportuj re
zaimportuj fnmatch

__all__ = ["glob", "iglob"]

def glob(pathname, *, recursive=Nieprawda):
    """Return a list of paths matching a pathname pattern.

    The pattern may contain simple shell-style wildcards a la
    fnmatch. However, unlike fnmatch, filenames starting przy a
    dot are special cases that are nie matched by '*' oraz '?'
    patterns.

    If recursive jest true, the pattern '**' will match any files oraz
    zero albo more directories oraz subdirectories.
    """
    zwróć list(iglob(pathname, recursive=recursive))

def iglob(pathname, *, recursive=Nieprawda):
    """Return an iterator which uzyskajs the paths matching a pathname pattern.

    The pattern may contain simple shell-style wildcards a la
    fnmatch. However, unlike fnmatch, filenames starting przy a
    dot are special cases that are nie matched by '*' oraz '?'
    patterns.

    If recursive jest true, the pattern '**' will match any files oraz
    zero albo more directories oraz subdirectories.
    """
    dirname, basename = os.path.split(pathname)
    jeżeli nie has_magic(pathname):
        jeżeli basename:
            jeżeli os.path.lexists(pathname):
                uzyskaj pathname
        inaczej:
            # Patterns ending przy a slash should match only directories
            jeżeli os.path.isdir(dirname):
                uzyskaj pathname
        zwróć
    jeżeli nie dirname:
        jeżeli recursive oraz _isrecursive(basename):
            uzyskaj z glob2(dirname, basename)
        inaczej:
            uzyskaj z glob1(dirname, basename)
        zwróć
    # `os.path.split()` returns the argument itself jako a dirname jeżeli it jest a
    # drive albo UNC path.  Prevent an infinite recursion jeżeli a drive albo UNC path
    # contains magic characters (i.e. r'\\?\C:').
    jeżeli dirname != pathname oraz has_magic(dirname):
        dirs = iglob(dirname, recursive=recursive)
    inaczej:
        dirs = [dirname]
    jeżeli has_magic(basename):
        jeżeli recursive oraz _isrecursive(basename):
            glob_in_dir = glob2
        inaczej:
            glob_in_dir = glob1
    inaczej:
        glob_in_dir = glob0
    dla dirname w dirs:
        dla name w glob_in_dir(dirname, basename):
            uzyskaj os.path.join(dirname, name)

# These 2 helper functions non-recursively glob inside a literal directory.
# They zwróć a list of basenames. `glob1` accepts a pattern dopóki `glob0`
# takes a literal basename (so it only has to check dla its existence).

def glob1(dirname, pattern):
    jeżeli nie dirname:
        jeżeli isinstance(pattern, bytes):
            dirname = bytes(os.curdir, 'ASCII')
        inaczej:
            dirname = os.curdir
    spróbuj:
        names = os.listdir(dirname)
    wyjąwszy OSError:
        zwróć []
    jeżeli nie _ishidden(pattern):
        names = [x dla x w names jeżeli nie _ishidden(x)]
    zwróć fnmatch.filter(names, pattern)

def glob0(dirname, basename):
    jeżeli nie basename:
        # `os.path.split()` returns an empty basename dla paths ending przy a
        # directory separator.  'q*x/' should match only directories.
        jeżeli os.path.isdir(dirname):
            zwróć [basename]
    inaczej:
        jeżeli os.path.lexists(os.path.join(dirname, basename)):
            zwróć [basename]
    zwróć []

# This helper function recursively uzyskajs relative pathnames inside a literal
# directory.

def glob2(dirname, pattern):
    assert _isrecursive(pattern)
    jeżeli dirname:
        uzyskaj pattern[:0]
    uzyskaj z _rlistdir(dirname)

# Recursively uzyskajs relative pathnames inside a literal directory.

def _rlistdir(dirname):
    jeżeli nie dirname:
        jeżeli isinstance(dirname, bytes):
            dirname = bytes(os.curdir, 'ASCII')
        inaczej:
            dirname = os.curdir
    spróbuj:
        names = os.listdir(dirname)
    wyjąwszy os.error:
        zwróć
    dla x w names:
        jeżeli nie _ishidden(x):
            uzyskaj x
            path = os.path.join(dirname, x) jeżeli dirname inaczej x
            dla y w _rlistdir(path):
                uzyskaj os.path.join(x, y)


magic_check = re.compile('([*?[])')
magic_check_bytes = re.compile(b'([*?[])')

def has_magic(s):
    jeżeli isinstance(s, bytes):
        match = magic_check_bytes.search(s)
    inaczej:
        match = magic_check.search(s)
    zwróć match jest nie Nic

def _ishidden(path):
    zwróć path[0] w ('.', b'.'[0])

def _isrecursive(pattern):
    jeżeli isinstance(pattern, bytes):
        zwróć pattern == b'**'
    inaczej:
        zwróć pattern == '**'

def escape(pathname):
    """Escape all special characters.
    """
    # Escaping jest done by wrapping any of "*?[" between square brackets.
    # Metacharacters do nie work w the drive part oraz shouldn't be escaped.
    drive, pathname = os.path.splitdrive(pathname)
    jeżeli isinstance(pathname, bytes):
        pathname = magic_check_bytes.sub(br'[\1]', pathname)
    inaczej:
        pathname = magic_check.sub(r'[\1]', pathname)
    zwróć drive + pathname
